Auth\Basic: add input validation to constructor
The `Basic::__construct()` method expects to receive an array of exactly two element and did validate the input received, but only threw an exception when the passed parameter did not match the expected element count, not when the passed parameter wasn't an array.

At the same time, the default value of `$args` is `null`, not an empty array and the `$user` and `$pass` properties are both public.

The only reason I can think of to have these properties `public` and the default value set to `null`, is to allow for delayed setting of the `$user` and `$pass`, after the class has already been instantiated.

To maintain the existing behaviour, but still add input validation, an exception is now thrown when the `$args` parameter is neither an array, nor `null`.

Includes;
* A dedicated test for the input validation.
* An additional test with Basic authentication with an instantiated class, but with delayed setting of the `$user` and `$pass` directly on the instantiated object. This test is solely added to document that this is a supported feature (and to safeguard against regressions in that respect).